Add Container Patch URLS at build time
Adds apply_patches.sh to keep the Dockerfile from growing too complex.
Can be used in `entrypoint.sh` as well I imagine.

Changes the optional build step to write a `status.txt` out. Later use that file to
know whether or not foundry was installed. I was unable to
find a way to detect this built into to Docker and multistage builds do
not persist environment variables across them. I am very open to other options.
Then read back the file to determine whether or not to run the Container
Patch URLs then remove it..

Container Patch URLs must be ran in the second stage because they may
rely on the `/data` directory existing and what not. Running in the
second step guarantees the same runtime environment they normally expect.